+++ collective/analyticspanel/testing.py
@@ -1,41 +1,49 @@
 # -*- coding: utf-8 -*-
-
-from zope.configuration import xmlconfig
-
-from plone.testing import z2
-
-from plone.app.testing import PLONE_FIXTURE
-from plone.app.testing import PloneSandboxLayer
-from plone.app.testing import IntegrationTesting
-from plone.app.testing import FunctionalTesting
+from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
 from plone.app.testing import applyProfile
-from plone.app.testing import setRoles
-from plone.app.testing import TEST_USER_ID
+from plone.app.testing import FunctionalTesting
+from plone.app.testing import IntegrationTesting
+from plone.app.testing import PloneSandboxLayer
+from zope.component import queryUtility
+from plone.registry.interfaces import IRegistry
+from plone import api
+try:
+    from Products.CMFPlone.interfaces.controlpanel import ISiteSchema
+    PLONE_4 = False
+except ImportError:
+    PLONE_4 = True
 
-class AnalyticsPanel(PloneSandboxLayer):
+import collective.analyticspanel
 
-    defaultBases = (PLONE_FIXTURE, )
+
+class AnalyticsPanelLayer(PloneSandboxLayer):
+
+    defaultBases = (PLONE_APP_CONTENTTYPES_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
-        # Load ZCML for this package
-        import collective.analyticspanel
-        xmlconfig.file('configure.zcml',
-                       collective.analyticspanel,
-                       context=configurationContext)
-        z2.installProduct(app, 'collective.analyticspanel')
+        self.loadZCML(package=collective.analyticspanel)
 
     def setUpPloneSite(self, portal):
-        portal.portal_properties.site_properties.webstats_js = 'SITE DEFAULT ANALYTICS'
+        if PLONE_4:
+            portal.portal_properties.site_properties.webstats_js = 'SITE DEFAULT ANALYTICS'
+        else:
+            registry = queryUtility(IRegistry)
+            site = registry.forInterface(ISiteSchema, prefix="plone")
+            site.webstats_js = u'SITE DEFAULT ANALYTICS'
+
         applyProfile(portal, 'collective.analyticspanel:default')
-        #quickInstallProduct(portal, 'collective.analyticspanel')
-        setRoles(portal, TEST_USER_ID, ['Member', 'Manager'])
 
 
-ANALYTICS_PANEL_FIXTURE = AnalyticsPanel()
-ANALYTICS_PANEL_INTEGRATION_TESTING = \
-    IntegrationTesting(bases=(ANALYTICS_PANEL_FIXTURE, ),
-                       name="AnalyticsPanel:Integration")
-ANALYTICS_PANEL_FUNCTIONAL_TESTING = \
-    FunctionalTesting(bases=(ANALYTICS_PANEL_FIXTURE, ),
-                       name="AnalyticsPanel:Functional")
+ANALYTICS_PANEL_FIXTURE = AnalyticsPanelLayer()
 
+
+ANALYTICS_PANEL_INTEGRATION_TESTING = IntegrationTesting(
+    bases=(ANALYTICS_PANEL_FIXTURE,),
+    name='AnalyticsPanelLayer:IntegrationTesting'
+)
+
+
+ANALYTICS_PANEL_FUNCTIONAL_TESTING = FunctionalTesting(
+    bases=(ANALYTICS_PANEL_FIXTURE,),
+    name='AnalyticsPanelLayer:FunctionalTesting'
+)