+++ collective/analyticspanel/tests/test_viewlet.py
@@ -1,38 +1,38 @@
 # -*- coding: utf-8 -*-
-
+from base import BaseTestCase
+from collective.analyticspanel.browser.viewlet import AnalyticsViewlet
+from collective.analyticspanel.pair_fields import ErrorCodeValuePair, SitePathValuePair
+from collective.analyticspanel.testing import ANALYTICS_PANEL_INTEGRATION_TESTING
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
 from zope.component import getMultiAdapter
 
-from collective.analyticspanel.testing import ANALYTICS_PANEL_INTEGRATION_TESTING
-from collective.analyticspanel.pair_fields import ErrorCodeValuePair, SitePathValuePair
-
-from collective.analyticspanel.browser.viewlet import AnalyticsViewlet
-
-from base import BaseTestCase
 
 class TestViewlet(BaseTestCase):
 
     layer = ANALYTICS_PANEL_INTEGRATION_TESTING
 
+    def setUp(self):
+        """Custom shared utility setup for tests."""
+        self.portal = self.layer['portal']
+        self.request = self.layer['request']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+
     def test_viewlet_registered(self):
-        portal = self.layer['portal']
-        request = self.layer['request']
         self.markRequestWithLayer()
-        request.set('ACTUAL_URL', 'http://nohost/plone')
+        self.request.set('ACTUAL_URL', 'http://nohost/plone')
         self.getSettings().general_code = u'SITE ANALYTICS FOOTER'
         self.getSettings().general_header_code = u'SITE ANALYTICS HEADER'
-        self.assertTrue('SITE ANALYTICS FOOTER' in portal())
-        self.assertTrue('SITE ANALYTICS HEADER' in portal())
+        self.assertTrue('SITE ANALYTICS FOOTER' in self.portal())
+        self.assertTrue('SITE ANALYTICS HEADER' in self.portal())
 
 # Do not run this test until p.a.testing will not fix https://dev.plone.org/ticket/11673
 #    def test_back_base_viewlet(self):
-#        portal = self.layer['portal']
 #        applyProfile(portal, 'collective.analyticspanel:uninstall')
 #        self.assertTrue('SITE DEFAULT ANALYTICS' in portal())
 
     def test_not_found(self):
-        portal = self.layer['portal']
-        request = self.layer['request']
-        request.set('ACTUAL_URL', 'http://nohost/plone')
+        self.request.set('ACTUAL_URL', 'http://nohost/plone')
         self.markRequestWithLayer()
         settings = self.getSettings()
 
@@ -47,16 +47,14 @@ class TestViewlet(BaseTestCase):
         record.position = 'header'
         settings.error_specific_code += (record,)
 
-        request.set('error_type', 'NotFound')
-        view = getMultiAdapter((portal, request), name=u"sharing")
+        self.request.set('error_type', 'NotFound')
+        view = getMultiAdapter((self.portal, self.request), name=u"sharing")
         self.assertTrue('You are in a NotFound page (footer)' in view())
         self.assertTrue('You are in a NotFound page (header)' in view())
 
     def test_for_path(self):
-        portal = self.layer['portal']
-        portal.invokeFactory(type_name='Folder', id='news', title="News")
-        request = self.layer['request']
-        request.set('ACTUAL_URL', 'http://nohost/plone/news')
+        self.portal.invokeFactory(type_name='Folder', id='news', title="News")
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news')
         self.markRequestWithLayer()
         settings = self.getSettings()
 
@@ -73,18 +71,16 @@ class TestViewlet(BaseTestCase):
         record.position = u'header'
         settings.path_specific_code += (record,)
 
-        self.assertTrue('You are in the News section (footer)' in portal.news())
-        self.assertTrue('You are in the News section (header)' in portal.news())
-        request.set('ACTUAL_URL', 'http://nohost/plone')
-        self.assertFalse('You are in the News section' in portal())
+        self.assertTrue('You are in the News section (footer)' in self.portal.news())
+        self.assertTrue('You are in the News section (header)' in self.portal.news())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone')
+        self.assertFalse('You are in the News section' in self.portal())
 
     def test_most_specific_path_used(self):
-        portal = self.layer['portal']
-        portal.invokeFactory(type_name='Folder', id='news', title="News")
-        portal.news.invokeFactory(type_name='Folder', id='subnews', title="Subfolder inside news")
+        self.portal.invokeFactory(type_name='Folder', id='news', title="News")
+        self.portal.news.invokeFactory(type_name='Folder', id='subnews', title="Subfolder inside news")
 
-        request = self.layer['request']
-        request.set('ACTUAL_URL', 'http://nohost/plone/news')
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news')
         self.markRequestWithLayer()
         settings = self.getSettings()
 
@@ -103,22 +99,20 @@ class TestViewlet(BaseTestCase):
         record3.position = u'header'
         settings.path_specific_code += (record1, record2, record3)
 
-        request.set('ACTUAL_URL', 'http://nohost/plone/news')
-        self.assertTrue('You are in the Subnews section (footer)' in portal.news.subnews())
-        self.assertTrue('You are in the News section (header)' in portal.news.subnews())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news')
+        self.assertTrue('You are in the Subnews section (footer)' in self.portal.news.subnews())
+        self.assertTrue('You are in the News section (header)' in self.portal.news.subnews())
 
     def test_hiding_code(self):
-        portal = self.layer['portal']
-        portal.invokeFactory(type_name='Folder', id='news', title="News")
+        self.portal.invokeFactory(type_name='Folder', id='news', title="News")
 
-        request = self.layer['request']
-        request.set('ACTUAL_URL', 'http://nohost/plone/news')
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news')
         self.markRequestWithLayer()
         settings = self.getSettings()
         settings.general_header_code = u'DEFAULT ANALYTICS IN HEADER'
 
-        self.assertTrue('SITE DEFAULT ANALYTICS' in portal.news())
-        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in portal.news())
+        self.assertTrue('SITE DEFAULT ANALYTICS' in self.portal.news())
+        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in self.portal.news())
 
         record = SitePathValuePair()
         record.path = u'/news'
@@ -132,21 +126,19 @@ class TestViewlet(BaseTestCase):
         record.position = 'header'
         settings.path_specific_code += (record,)
 
-        self.assertFalse('SITE DEFAULT ANALYTICS' in portal.news())
-        self.assertFalse('DEFAULT ANALYTICS IN HEADER' in portal.news())
+        self.assertFalse('SITE DEFAULT ANALYTICS' in self.portal.news())
+        self.assertFalse('DEFAULT ANALYTICS IN HEADER' in self.portal.news())
 
     def test_no_subtree_propagation(self):
-        portal = self.layer['portal']
-        portal.invokeFactory(type_name='Folder', id='news', title="News")
+        self.portal.invokeFactory(type_name='Folder', id='news', title="News")
 
-        request = self.layer['request']
-        request.set('ACTUAL_URL', 'http://nohost/plone/news')
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news')
         self.markRequestWithLayer()
         settings = self.getSettings()
         settings.general_header_code = u'DEFAULT ANALYTICS IN HEADER'
 
-        self.assertTrue('SITE DEFAULT ANALYTICS' in portal.news())
-        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in portal.news())
+        self.assertTrue('SITE DEFAULT ANALYTICS' in self.portal.news())
+        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in self.portal.news())
 
         record = SitePathValuePair()
         record.path = u'/news'
@@ -160,26 +152,25 @@ class TestViewlet(BaseTestCase):
         record.position = u'header'
         settings.path_specific_code += (record,)
 
-        self.assertTrue('Only for news (footer)' in portal.news())
-        self.assertTrue('Only for news (header)' in portal.news())
-        
-        portal.news.invokeFactory(type_name='Folder', id='subnews', title="Subnews")
-        request.set('ACTUAL_URL', 'http://nohost/plone/news/subnews')
+        self.assertTrue('Only for news (footer)' in self.portal.news())
+        self.assertTrue('Only for news (header)' in self.portal.news())
 
-        self.assertTrue('SITE DEFAULT ANALYTICS' in portal.news.subnews())
-        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in portal.news.subnews())
+        self.portal.news.invokeFactory(
+            type_name='Folder', id='subnews', title="Subnews")
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news/subnews')
+
+        self.assertTrue('SITE DEFAULT ANALYTICS' in self.portal.news.subnews())
+        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in self.portal.news.subnews())
 
     def test_apply_to_folder_and_children(self):
-        request = self.layer['request']
         self.markRequestWithLayer()
         settings = self.getSettings()
         settings.general_header_code = u'DEFAULT ANALYTICS IN HEADER'
-
-        portal = self.layer['portal']
-        portal.invokeFactory(type_name='Folder', id='news', title="News")
-        news = portal.news
+        self.portal.invokeFactory(type_name='Folder', id='news', title="News")
+        news = self.portal.news
         news.invokeFactory(type_name='Folder', id='subnews', title="Sub news")
-        news.invokeFactory(type_name='Document', id='home', title="Homepage for news section")
+        news.invokeFactory(
+            type_name='Document', id='home', title="Homepage for news section")
 
         record = SitePathValuePair()
         record.path = u'/news'
@@ -193,25 +184,23 @@ class TestViewlet(BaseTestCase):
         record.position = u'header'
         settings.path_specific_code += (record,)
 
-        request.set('ACTUAL_URL', 'http://nohost/plone/news')
-        self.assertTrue('For news and children (footer)' in portal.news())
-        self.assertTrue('For news and children (header)' in portal.news())
-        request.set('ACTUAL_URL', 'http://nohost/plone/news/home')
-        self.assertTrue('For news and children (footer)' in portal.news.home())
-        self.assertTrue('For news and children (header)' in portal.news.home())
-        request.set('ACTUAL_URL', 'http://nohost/plone/news/subnews')
-        self.assertTrue('SITE DEFAULT ANALYTICS' in portal.news.subnews())
-        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in portal.news.subnews())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news')
+        self.assertTrue('For news and children (footer)' in self.portal.news())
+        self.assertTrue('For news and children (header)' in self.portal.news())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news/home')
+        self.assertTrue('For news and children (footer)' in self.portal.news.home())
+        self.assertTrue('For news and children (header)' in self.portal.news.home())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news/subnews')
+        self.assertTrue('SITE DEFAULT ANALYTICS' in self.portal.news.subnews())
+        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in self.portal.news.subnews())
 
     def test_apply_to_folder_and_children_with_new_folderish(self):
-        request = self.layer['request']
         self.markRequestWithLayer()
         settings = self.getSettings()
         settings.general_header_code = u'DEFAULT ANALYTICS IN HEADER'
 
-        portal = self.layer['portal']
-        portal.invokeFactory(type_name='Folder', id='news', title="News")
-        news = portal.news
+        self.portal.invokeFactory(type_name='Folder', id='news', title="News")
+        news = self.portal.news
         news.invokeFactory(type_name='News Item', id='real_news', title="A real news")
         news.invokeFactory(type_name='Document', id='home', title="Homepage for news section")
 
@@ -229,35 +218,36 @@ class TestViewlet(BaseTestCase):
         settings.path_specific_code += (record,)
         settings.folderish_types = (u'Folder', u'News Item')
 
-        request.set('ACTUAL_URL', 'http://nohost/plone/news')
-        self.assertTrue('For news and children (footer)' in portal.news())
-        self.assertTrue('For news and children (header)' in portal.news())
-        request.set('ACTUAL_URL', 'http://nohost/plone/news/home')
-        self.assertTrue('For news and children (footer)' in portal.news.home())
-        self.assertTrue('For news and children (header)' in portal.news.home())
-        request.set('ACTUAL_URL', 'http://nohost/plone/news/real_news')
-        self.assertTrue('SITE DEFAULT ANALYTICS' in portal.news.real_news())
-        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in portal.news.real_news())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news')
+        self.assertTrue('For news and children (footer)' in self.portal.news())
+        self.assertTrue('For news and children (header)' in self.portal.news())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news/home')
+        self.assertTrue('For news and children (footer)' in self.portal.news.home())
+        self.assertTrue('For news and children (header)' in self.portal.news.home())
+        self.request.set('ACTUAL_URL', 'http://nohost/plone/news/real_news')
+        self.assertTrue('SITE DEFAULT ANALYTICS' in self.portal.news.real_news())
+        self.assertTrue('DEFAULT ANALYTICS IN HEADER' in self.portal.news.real_news())
 
     def test_optout(self):
-        request = self.layer['request']
-        portal = self.layer['portal']
         self.markRequestWithLayer()
         settings = self.getSettings()
         settings.general_header_code = u'DEFAULT ANALYTICS IN HEADER'
         settings.respect_optout = True
-        request.cookies = {'analytics-optout': 'true'}
-        self.assertFalse('DEFAULT ANALYTICS IN HEADER' in portal())
+        self.request.cookies = {'analytics-optout': 'true'}
+        self.assertFalse('DEFAULT ANALYTICS IN HEADER' in self.portal())
 
 
 class TestViewletPathCleanup(BaseTestCase):
-    
+
     layer = ANALYTICS_PANEL_INTEGRATION_TESTING
-    
+
+    def setUp(self):
+        """Custom shared utility setup for tests."""
+        self.portal = self.layer['portal']
+        self.request = self.layer['request']
+
     def getViewlet(self):
-        portal = self.layer['portal']
-        request = self.layer['request']        
-        return AnalyticsViewlet(portal, request, None, None)
+        return AnalyticsViewlet(self.portal, self.request, None, None)
 
     def test_add_starting_slash(self):
         viewlet = self.getViewlet()